USD: cleanup and refactor domelight IO
Now creating a mapping node to set the environment texture
rotations.

No longer creating empties for dome lights on import, to reduce
clutter in the scene and reduce the chance of name collisions
for the dome light prims on export on round-trips.

Added logic to ensure the dome light paths are unique in the
scene.

Removed code that was not essential to dome lights to expedite
the review process.

Moved duplicated to code to separate functios.

General cleanup (e.g., replacing cerr/cout with CLOG).