Anim: add 'Frame Scene Range' to dope sheet & graph editor pie menu
Add a new operator 'Frame Scene Range' to the Dope Sheet and Graph Editor
'view' menus. It is added both to the regular menu and the pie menu (by
default on the ` hotkey).

The operator will only change the horizontal view, to show the scene range
or the preview range, depending on whether the preview range is active
or not.