GPv3: Refactor: Use GPv2 point times as is
During the conversion of GPv3 to GPv2, we computed the
time difference between the points. So we we're keeping track
of the time a point was created after the point before it.

There doesn't seem to be a good reason to do this.
It also can go very wrong in case the times of two points are
very close to one another (which can result in catastrophic
cancellation).