Fix: Assert when calling render operator
Blender asserts and behaves unpredictably when calling the render
operator from a python script when the GPU compositor was previously
used interactively in the node editor. That's due to the following.

The render pipeline assumes the system GPU context is nullptr when doing
a blocking render, that is, a render from the main thread. And while the
render pipeline cleanups after every render to ensure a nullptr system
GPU context, the GPU compositor initializes the system GPU context once
and doesn't discard it to improve interactive editing, since context
creation is relatively expensive. So we need to ensure a nullptr system
GPU context here before executing the render.