WIP: RNA: Add info about chain of ancestors (owner data) of a PointerRNA.
Disclaimer: This is not even WIP, more like a proof of concept. A
complete, fully implementation of this idea would require quite some
more work.

The general idea is to store an array of (type, data) pointers of all
PointerRNA ancestors of the current one.

This would help solving cases in our code where the owner (or someties
even the owner of the owner) of a random PointerRNA needs to be
accessed. Current solution mainly relies on linear search from the owner
ID, which is sub-optimal at best, and may not even be possible in case a
same data is shared between different owners.