Extensions: Don't print errors in console for automatic checks
The problem is that when a computer is offline, Blender will now show
console errors when it starts. We can't distinguish between the case
where this is expected and where it really is an error, so don't show
anything.

For troubleshooting, the user can still go to the preferences and
manually click check for updates. Then it will show the error.