Cleanup: Remove redundant is_viewer_socket function
A new is_viewer_socket function was copied from Node Wrangler after it
was ported to Blender core. In NW, it checked for a socket's name. But
it also checks the is_inspect_output property, which is only used for
this purpose. This makes the name check redundant, and the whole
function as well.
This change removes the function.