GPv3: Fill tool: Don't generate outlines for holes.
The fill tool generated strokes for exterior as well as interior
boundaries. This is undesirable because fill strokes can't actually have
holes and the overlapping strokes are not very useful.

This reconstructs GPv2 behavior where only the outer-most starting
points for boundaries are used, ignoring interior boundaries.

Also includes a fix to prevent crashes when
`curves_copy_point_selection` is called with an empty selection.