Nodes: use actual socket type as tmp viewer for Connect to Output
The Connect to Output operator is used to inspect sockets. If the node
to preview is inside a group, a temporary viewer socket is created.

Currently this uses hardcoded socket types (Geometry for Geometry
Nodes, Shader for shader nodes). While in GN we can only connect
geometry to the output, in shader nodes other types can be inspected.
This change allows the tmp_viewer to use the type of the actual socket
being inspected, to have a better idea of what it contains from
outside the group.

It can be especially useful if a group is used in multiple materials
and different sockets are being previewed in each one.