Fix #122466: Underlines on Text objects have overlapping geometry
It has been like that since the introduction in a07394ef2c.

There is a hardcoded overlap of 0.1 , see `uloverlap = xtrax + 0.1f;`

Would propose to just remove it, dont think this would actually cause
more problems than this z-fighting.

Underlines are separate per character, so whenever text is deformed,
this would break anyways, and for "regular" usage, there does not seem
to be a point in having this "artificial" overlap. The only thing that
would break earlier is having a negative `Geometry` `Offset`.