Fix #122438: Orbit Around Selection drastically moves view
Would happen if an object that is out of view is selected and then
navigating.

Caused by 172221e7fb89.

This part of the code has gone though some refactors:
- 15a4609ecefc
- 6faa39edb7
- 384c2e1f36eb

What we can notice here is that part of the code that used to be
reserved for `VIEWOPS_FLAG_DEPTH_NAVIGATE` was now called for different
conditions:
- in 6faa39edb7, it was now depending on `V3D_OP_MODE_ROTATE` (without
further explanation), caused #114714
- in 172221e7fb89 (the fix for #114714), it was now done unconditionally
(the explanation is not clear to me either)

To resolve, make the specific part of the code depend on
`VIEWOPS_FLAG_DEPTH_NAVIGATE` again