Refactor: GPv3: Move stroke operation construction in `stroke_update_step`
Previously, the stroke operation were created in the `stroke_test_start`
callback function. This causes a few issues:
* Each mode has it's own `test_start` function and it's own function
  to create the stroke operations.
* Only the mouse position was availible for the start sample.

With this change, the `stroke_test_start` function now does nothing.
The operations are created in the common `stroke_update_step`
function which calls `get_stroke_operation`.
This also means that the start sample can now get access to the
start pressure.