Anim: use fixed margin for Timeline / Dope Sheet / Graph Editor
The 'Frame All', 'Frame Selected', and 'Frame Scene/Preview Range'
operators will now use a fixed margin (in UI-scaled pixels). Previously
the margin was relative, and always used 10% of the available editor
width. This was considered too large, especially when the editor was
wide.

This unifies the margin calculations for the graph editor, dope sheet,
and timeline. Earlier versions used smaller margins for the graph editor
than for the other editors, but this is no longer necessary.