Node Editor: Always allow inserting reroute nodes
Since ebb61ef30f reroute nodes can't be inserted on links anymore,
when their sockets currently have an incompatible type.

Because the sockets on reroute nodes change types dynamically
according to the sockets they are linked to, this commit adds an
exception to always allow reroutes to be inserted onto links.