Cleanup: Removing unused or non-working parts.
This also removes existing code for join_geometries and
realize_instances for the PhysicsComponent. The physics data is not
trivially copyable (comes with big overhead) and there isn't a good
design for using split/join flows in geometry nodes for this yet.

For the time being nodes will be modified to take an extra mutable
input with physics geometry, avoiding copies.