Allow storing multiple "impls" (worlds?) per physics geometry.
The purpose of this somewhat convoluted data model is to enable geometry
nodes workflow with Join nodes and realizing components. This requires
that the GeometryComponent data is copyable, since the source geometries
are immutable during the operation.

Physics world data (Bullet) are quite large (700b per body), and there
is no native support for deep copying. For simulation workflow,
not being able to branch the data is acceptable, but it should support
idioms like the Join Geometries node and simulation zones.

The idea being tested here is to copy only an implicit shared pointer to
the physics data, and then "consolidate" the data into a single world
once needed.