UI: Fix polling in sequencer preview overlays
When the sequencer overlays menu was restructured in #121591, the new
Sequencer panel types did not include polling, which means that they
also show up in the Preview-only view type. This PR polls them so that
overlays are properly divided again.

To accommodate the wide panel, this patch also organizes preview
overlays into a two-column structure like sequencer overlays.

With this change, popover menu in Preview is totally blank when choosing
display mode other than image (which was the behavior in previous
blender versions), so to make things less confusing for the user, make
these options inactive rather than hidden to indicate that changing them
will have no effect in the current mode.