EEVEE-Next: Crash when using sculpting template
The sculpting template was created in Blender 2.80 and never adjusted
since then. There is no version code for adding the Alpha socket
to the principled BSDF.

Although we should add versioning code to add the alpha socket this
Fix only ensures that the recent added versioning code will not crash.