Cycles: Fix assert failure for CPU render with no CPU device selected
This is a proper fix for the issue worked around in 11d311e3007b.

Previously, default-initialized DeviceInfo would pretend being
a device with "CPU" id, which would cause operator==() to fail
the assert while comparing to any real CPU device due to
differences in other fields.

With this change, default-initialized DeviceInfo will have
"None" as id, clearly indicating, that this is a Null object
for DeviceInfo class instead of a CPU device.