Refactor: Attributes API: Remove dependency on ID
The attribute API defined in `attribute.cc` was dependent on
the assumption that `ID`s are always the "direct" owner of attributes.

For Grease Pencil drawings, this is not the case.
In order to make use of  `rna_attribute.cc`, we need that API to handle
other types of attribute owners.

This adds an `AttributeOwner` which is basically just a type and a
pointer. We replace the `ID` pointers and pass `AttributeOwner`s instead.