Cleanup: Sculpt: Remove object idname from sculpt undo node
In preparation for improving the performance of this area I'm trying
to simplify it a bit and fix some confusing things that make the area
buggy or hard to work with and buggy.

One of these it that each node stores the object's name. This is
unnecessary because sculpt mode only works on a single object.
But even if it did, storing that on every node doesn't make sense,
there might be hundreds stored in a single undo step. If multi-
object sculpting is added in the future, we would have a separate
structure of nodes per object.