Refactor: Anim: make some `bAnimListElem` fields enum-typed
Change the `bAnimListElem` fields `type`, `update`, and `datatype` from
respectively `int`, `char`, and `short` to their actual `enum` types,
respectively `eAnim_ChannelType`, `eAnim_Update_Flags`, and `eAnim_KeyType`.

To prevent compiler warnings, all `switch` statements that try to handle
these fields have been expanded with the missing `case`s. This should help
in the future when new channel types are added, as that'll cause more
compiler warnings in those places that need updating.

One `if`/`else if`/`else` chain was converted into a `switch` for clarity.
This was actually the place where my (upcoming) code is hitting the
`BLI_assert_unreachable()`, which was the final straw for this refactor.

No functional changes.