Refactor: Use C++ types to manage custom UI block interaction data
Was doing manual (error prone, verbose and DRY violating) array and user
count management. Use `blender::Set` and `std::shared_ptr` instead.

Small annoyance is that this makes `uiHandleButtonData` non-trivial to
construct, so I had to change allocate functions. But was doing this in
another refactor that prompted me to do this one anyway.