Anim: API change: remove the RNA enum property `AnimData.action_binding`
Remove the `animation_data.action_binding` enum property. In an upcoming
commit it will become a pointer property instead, which can then be used to
get or set the assigned action binding.

Having this as a remove-only commit, to be followed by an add-only commit,
makes the diff significantly easier to understand.

The bindings are temporarily removed from the developer-oriented GUI to
prevent runtime errors.