Anim: add Action + Binding selector to the Camera properties editor
Add an Action selector + a Binding selector to the Camera properties
editor.

This makes it possible to switch the Camera data-block to another Action
(legacy or layered). If a layered Action is assigned, another selector is
shown to assign a Binding.

There are no operators yet for creating a new Binding or for unassigning
the assigned Binding.