Compositor: Make Auto Render work on all object changes
Currently, the Auto render option works by executing the auto render
pipeline when an object is transformed or when playback stops. This
patch changes the mechanism to make it execute whenever any object in
the scene changes directly or indirectly. This is done by tagging the
compositor inside depsgraph editor scene updates if a change to an
object is detected.