UI: Separate job type for asset library and file list loading
Asset library loading uses the file browser backend (file-list), so it
would use the same job type. The job system makes sure that jobs of the
same type (or actually, with the same start callback before this PR)
wait for others of the same type to finish. This can be a problem here,
since loading asset libraries (which can take a while) could conflict
with regular file browers. Having both run in parallel is no issue, they
use local data only.

Also changes the jobs system to use the job type for suspending alike
jobs, rather than the start callback. This is required for this to work,
but will submit a separate PR.

Fixes #121235