Refactor: WM: Suspend new jobs based on job type, not start callback
The logic here was introduced in b466286c3e, before job types were added
in 7b60529517. It seems like the job type is the more predictable/sane
way to identify jobs that should be exclusive. Was surprised that a
separate job type wouldn't be enough to separate jobs when working on
 #123027.

Checked all existing jobs to make sure behavior is unchanged. Found
two issues:
- `WM_JOB_TYPE_OBJECT_SIM_FLUID` is used for both
  `fluid_bake_startjob()` and `fluid_free_startjob()`. It makes sense to
  me that they would be exclusive though, so leaving it this way
  (meaning they are exclusive now).
- Alembic and USD job types were reused, split them up now to not change
  behavior.