Fix: UI: make statistics for Objects display 0 instead of (null)
When there are no objects in the scene, the viewport statisticts
currently display:

Objects     (null)

This is because an int with value of 0 is passed to the formatting
function in `stats_row`.

Since the number of objects is gathered in stats_fmt as a string, it
can be used instead of the hardcoded 0 value.